Emulator core is now entirely cross platform
Some code deep in the emulator depended on the host system to have a
little-endian byte order. This commit removes this requirement and, in
the process, makes some code much easier to read and think about.

This might lead to very small performance regressions, but this is
probably a price worth paying.